requirements.txt
pytesseract
pillow
pdf2image
tqdm
pandas
opencv-python
streamlit
